rename DbAdapter
